+++ examples/resources/catalystcenter_planned_access_point_position/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_planned_access_point_position.example "<floor_id>,<name>"